♻ README : use docker-compose consistently
We suggest to run all commands through the Docker Compose CLI, in so we
have a stable reference for the way how we do things with the
containers.

Also manage.py is favoured over django-admin.py, as it is the official successor.

Note that the `-w WORKDIR` Docker Compose parameter is not supported on older Docker engines, why the
workaround with an enclosed bash command `sh -c 'cd ... && '` is being used.